renamed all the things, exclude firebase dir
