--- .firebase/hosting.YnVpbGQ.cache
@@ -1,20 +1,0 @@
-asset-manifest.json,1586906979024,730b50aeaf8b022108a1d42bc2aba770162754286435d08f3c64a8da8c2e6f11
-index.html,1586906979024,dbf1db7f95c2cd73a6365d8f97964b47b9297b51f4ac9561f20dc99d4eda97ef
-favicon.ico,499162500000,eae62e993eb980ec8a25058c39d5a51feab118bd2100c4deebb2a9c158ec11f9
-precache-manifest.3e13ce5b0c5fa1643b848b1d71e756c7.js,1586906979024,ce00b519227ed13fc82933ec468a28abcdc4e3e3193ed094c8f0dfff14a6232a
-manifest.json,499162500000,87b4297a1590fbfbf48bd5a77c4fcd77f7e1309c23abfdde25434891584c7e88
-logo192.png,499162500000,3ee59515172ee198f3be375979df15ac5345183e656720a381b8872b2a39dc8b
-robots.txt,499162500000,5fcd5c559ded0c02b3ed7840ca3ee77e95b798730af98d9f18bc627ac898071e
-service-worker.js,1586906979024,e9979ae8f4207b204928a8fb0676c1d88e08af03827e71ce80d13a4d693efba9
-logo512.png,499162500000,ee7e2f3fdb8209c4b6fd7bef6ba50d1b9dba30a25bb5c3126df057e1cb6f5331
-static/css/main.e0a69770.chunk.css,1586906979025,821dbe2a0d210c20bf26c94d2228ade6f7b67656e8e59ab4154b9a447e2b6f1b
-static/css/main.e0a69770.chunk.css.map,1586906979026,c4deffc9226cd8383436c8a1c99f786d5dbda4795b87a60e084ff034c4c30913
-static/js/2.a315fdce.chunk.js.LICENSE.txt,1586906979026,26ebefebef26088c3fcf10ef730b713d27d2b5aed26fda8cea225180f7c13e9d
-static/js/main.4e39cd5a.chunk.js,1586906979025,64b9c633860d4d6717c2711925f1d90d5112c78a9baa12481fc8ff9bfba36ebf
-static/js/main.4e39cd5a.chunk.js.map,1586906979026,f2b6bd31fd54756e3ed879424fec1c6ebd323aac3ddd9e742644e54dd24d8bfa
-static/js/runtime-main.d1594a52.js,1586906979026,acc5e496a7eed0ee6cd6bca04adaf4f1e52bdd5fc764ae94f02be985776c3810
-static/js/runtime-main.d1594a52.js.map,1586906979027,7fd8e79a9b996ec887d239388c3c23c28dc691c4c02d534e2f0b507db576699d
-static/css/2.89f067aa.chunk.css,1586906979026,600cf3c52ebfe26a90720e8f5d043457a66a792e400afa32c600934521218421
-static/js/2.a315fdce.chunk.js,1586906979026,24ed9d5350f2fa7b308154798315d483b0ce1cfa86698e3711a14a34f763ce14
-static/css/2.89f067aa.chunk.css.map,1586906979027,d7c7591a355f47a645d8f14e801f9a3bf6c848c1f4810cae56d189d09d3c295c
-static/js/2.a315fdce.chunk.js.map,1586906979027,1505af8e820e93db554669e45ba6e26d0081676bbf8e401f56c59df4ac491d50
