+++ src/index.js
@@ -3,7 +3,7 @@ import "bootstrap/dist/css/bootstrap.css";
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
+import App from "./App.jsx";
 import * as serviceWorker from "./serviceWorker";
 
 var text;
